+++ cj.netos.gateway.wallet.program/src/cj/netos/gateway/wallet/IDepositAbsorbActivityController.java
@@ -4,8 +4,10 @@ import cj.netos.gateway.wallet.model.DepositAbsorbRecord;
 import cj.netos.gateway.wallet.result.DepositAbsorbResult;
 import cj.studio.ecm.net.CircuitException;
 
+import java.math.BigDecimal;
+
 public interface IDepositAbsorbActivityController {
-    DepositAbsorbRecord doReceipt(String principal, String personName, long amount, String sourceCode, String sourceTitle, String note) throws CircuitException;
+    DepositAbsorbRecord doReceipt(String principal, String personName, BigDecimal amount, String sourceCode, String sourceTitle, String note) throws CircuitException;
 
     void ackReceipt(DepositAbsorbResult result);
